int32 ==> int16

+++ attnptrns/treemap_longformer_attn.py
@@ -45,7 +45,7 @@ class AttentionPattern():
     max_graph_len = max([receivers.shape[0] for receivers in receivers_heads])
     r, s, m = [], [], []
     def pad_to(mat, padding):
-      padded_mat = jnp.zeros((padding), dtype=jnp.int32)
+      padded_mat = jnp.zeros((padding), dtype=jnp.int16)
       padded_mat = padded_mat.at[:mat.shape[0]].set(mat)
       return padded_mat
     def get_mask(mat, padding):
